Add files via upload
Write a Python script that would read the `data.json` file and print the name of the actor, the number of movies they starred in, and the average rating of all the movies they starred in. Do not print actors that starred in less than 2 movies. The output should be printed in ascending order by the total number of movies.